+++ src/GUI/gui.py
@@ -49,10 +49,6 @@ class MainView(QMainWindow):
                 "session": {
                     "opened_folder": "",
                     "opened_file": ""
-                },
-                "credits": {
-                    "username": "",
-                    "password": ""
                 }
             }
             with open('config.json', 'w') as file: