+++ src/GUI/info_win.py
@@ -3,9 +3,12 @@ from PySide6.QtGui import QIcon, QFont, QColor
 from PySide6.QtCore import Qt
 import sys
 from Tools import scrap
+import json
 
 def get_current_version():
-    return "2.1.5"
+    with open("config.json", "r") as f:
+            config_data = json.load(f)    
+    return config_data.get("version")
 
 class VersionWindow(QWidget):
     def __init__(self, theme=None):